Start cleaning code (remove openid and cas, update server info endpoint)
