+++ twake/frontend/src/app/scenes/Client/Popup/CreateCompanyView/CreateCompanyView.js
@@ -323,7 +323,7 @@ export default class CreateCompanyView extends Component {
       );
     }
     if (this.state.page == 3) {
-      if (InitService.server_infos?.auth?.console?.use) {
+      if (InitService.server_infos?.configuration?.account?.type === 'console') {
         return (
           <AddUserFromTwakeConsole
             inline
@@ -405,8 +405,8 @@ export default class CreateCompanyView extends Component {
     } else {
       if (
         (['openid', 'cas'].indexOf((CurrentUser.get() || {}).identity_provider) >= 0 ||
-          (((InitService.server_infos || {}).auth || {}).internal || {})
-            .disable_email_verification) &&
+          InitService.server_infos?.configuration?.account?.type ===
+            'internal'.disable_email_verification) &&
         this.state.page == 3
       ) {
         this.state.page = 2;
@@ -434,8 +434,8 @@ export default class CreateCompanyView extends Component {
       //Pass usage form
       if (
         (['openid', 'cas'].indexOf((CurrentUser.get() || {}).identity_provider) >= 0 ||
-          (((InitService.server_infos || {}).auth || {}).internal || {})
-            .disable_email_verification) &&
+          InitService.server_infos?.configuration?.account?.type ===
+            'internal'.disable_email_verification) &&
         this.state.page == 1
       ) {
         this.state.page = 2;
