+++ twake/frontend/src/app/scenes/Client/Popup/CreateWorkspacePage/CreateWorkspacePage.js
@@ -84,11 +84,11 @@ export default class CreateWorkspacePage extends Component {
       );
     }
     if (this.state.page == 2) {
-      if (InitService.server_infos?.auth?.console?.use) {
+      if (InitService.server_infos?.configuration?.account?.type === 'console') {
         return (
           <AddUserFromTwakeConsole
             inline
-            onChange={(members) => {
+            onChange={members => {
               this.state.members = members;
               this.setState({});
             }}
