+++ twake/frontend/src/app/scenes/Client/Popup/WorkspaceParameter/Pages/WorkspacePartnerTabs/Members.js
@@ -24,7 +24,7 @@ export default class Members extends React.Component {
           onAdd={
             workspaceUserRightsService.hasWorkspacePrivilege() &&
             (() => {
-              if (InitService.server_infos?.auth?.console?.use) {
+              if (InitService.server_infos?.configuration?.account?.type === 'console') {
                 return popupManager.open(<AddUserFromTwakeConsole standalone />);
               } else {
                 return popupManager.open(<AddUser standalone />);
