+++ twake/frontend/src/app/scenes/Login/LoginView/LoginView.js
@@ -19,7 +19,7 @@ export default class LoginView extends Component {
     Languages.addListener(this);
   }
   componentDidMount() {
-    const authModes = Object.keys((InitService.server_infos || {}).auth || {});
+    const authModes = [InitService.server_infos?.configuration?.accounts?.type];
     if (
       authModes.length > 0 &&
       authModes.indexOf('internal') < 0 &&
@@ -35,7 +35,7 @@ export default class LoginView extends Component {
   render() {
     const login = this.state.login;
 
-    const authModes = Object.keys((InitService.server_infos || {}).auth || {});
+    const authModes = [InitService.server_infos?.configuration?.accounts?.type];
     if (
       authModes.length > 0 &&
       authModes.indexOf('internal') < 0 &&
@@ -63,36 +63,6 @@ export default class LoginView extends Component {
               style={{ marginBottom: 40, marginTop: 10, width: 140 }}
               src={((InitService.server_infos || {}).branding || {}).logo}
             />
-          )}
-
-          {Object.keys((InitService.server_infos || {}).auth || []).indexOf('cas') >= 0 && (
-            <div className="external-login" style={{ marginBottom: 16 }}>
-              <Button
-                id="login_btn"
-                type="button"
-                className="medium full_width "
-                style={{ marginBottom: 8 }}
-                disabled={this.state.login.login_loading}
-                onClick={() => LoginService.loginWithExternalProvider('cas')}
-              >
-                {this.state.i18n.t('scenes.login.home.login_external_btn', ['CAS'])}
-              </Button>
-            </div>
-          )}
-
-          {Object.keys((InitService.server_infos || {}).auth || []).indexOf('openid') >= 0 && (
-            <div className="external-login" style={{ marginBottom: 16 }}>
-              <Button
-                id="login_btn"
-                type="button"
-                className="medium full_width "
-                style={{ marginBottom: 8 }}
-                disabled={this.state.login.login_loading}
-                onClick={() => LoginService.loginWithExternalProvider('openid')}
-              >
-                {this.state.i18n.t('scenes.login.home.login_external_btn', ['OpenID'])}
-              </Button>
-            </div>
           )}
 
           {this.state.login.external_login_error && (
@@ -154,8 +124,8 @@ export default class LoginView extends Component {
               >
                 {this.state.i18n.t('scenes.login.home.login_btn')}
               </Button>
-              {!(((InitService.server_infos || {}).auth || {}).internal || {})
-                .disable_account_creation && (
+              {!InitService.server_infos?.configuration?.account?.type ===
+                'internal'.disable_account_creation && (
                 <a
                   onClick={() => this.state.login.changeState('signin')}
                   id="create_btn"
