+++ twake/frontend/src/app/scenes/Login/VerifyMail/VerifyMail.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 
 import Languages from 'services/languages/languages.js';
 import LoginService from 'services/login/login.js';
+import AccountService from 'services/login/account';
 import Emojione from 'components/Emojione/Emojione';
 import WindowState from 'services/utils/window.js';
 
@@ -18,7 +19,7 @@ export default class VerifyMail extends Component {
     Languages.addListener(this);
   }
   componentDidMount() {
-    LoginService.doVerifyMail(
+    AccountService.doVerifyMail(
       WindowState.findGetParameter('m'),
       WindowState.findGetParameter('c'),
       WindowState.findGetParameter('token'),
