--- twake/frontend/src/app/scenes/Setup/Setup.scss
@@ -1,72 +1,0 @@
-.centered-twake-not-ready-page {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-
-  .twake-not-ready-page {
-    text-align: center;
-    width: 600px;
-    padding: 40px;
-    margin: auto;
-
-    .title {
-      font-size: 40px;
-      .emojione {
-        margin-right: 5px;
-        width: 64px;
-        height: 64px;
-        bottom: 8px;
-      }
-    }
-    .subtitle {
-      text-align: center;
-      line-height: 25px;
-      margin: auto;
-      margin-top: 30px;
-      margin-bottom: 40px;
-      max-width: 500px;
-    }
-    .signed {
-      text-align: right;
-      opacity: 0.7;
-      font-size: 13px;
-    }
-    .btn {
-      height: 40px;
-      line-height: 40px;
-      border-radius: var(--border-radius-base);
-      display: inline-block;
-      padding: 0px 20px;
-    }
-    .blueBtn {
-      background: var(--primary);
-      color: white;
-    }
-    .btn:hover {
-      cursor: pointer;
-      box-shadow: inset 0px 0px 100px 100px rgba(255, 255, 255, 0.2);
-    }
-    .retry {
-      border: none;
-      margin-top: 40px;
-
-      .link {
-        color: #888;
-      }
-    }
-    .otherMail {
-      border: none;
-      margin-top: 40px;
-      line-height: 20px;
-
-      .label {
-        border: none;
-      }
-      .link {
-        border: none;
-      }
-    }
-  }
-}
