+++ twake/frontend/src/app/services/InitService.ts
@@ -1,45 +1,46 @@
 import Observable from 'services/Observable/Observable';
 import Api from 'services/Api';
-import Websocket from 'services/websocket.js';
-import RouterServices from 'app/services/RouterService';
-import { ConfigurationType } from 'app/models/Configuration';
 
-type ServerInfoType = {
-  branding?: any;
-  websocket_public_key?: string;
-  auth?: ConfigurationType['auth'];
-  ready?:
-    | true
-    | {
-        elasticsearch_connection: boolean;
-        elasticsearch_mapping: boolean;
-        db_connection: boolean;
-        db_mapping: boolean;
-        init: boolean;
+type ServerInfoType = null | {
+  status: 'ready';
+  version: {
+    current: string;
+    minimal: {
+      web: string;
+      mobile: string;
+    };
+  };
+  configuration: {
+    branding: any;
+    help_link: string | null;
+    accounts: {
+      type: 'console' | 'internal';
+      console: null | {
+        account_management_url: string;
+        company_management_url: string;
+        collaborators_management_url: string;
       };
-  help_link?: string;
+      internal: null | {
+        disable_account_creation: boolean;
+        disable_email_verification: boolean;
+      };
+    };
+  };
 };
 
 class InitService extends Observable {
-  public server_infos: ServerInfoType = {};
+  public server_infos: ServerInfoType = null;
   public server_infos_loaded: boolean = false;
   public app_ready: boolean = false;
 
   async init() {
-    const res: any = await Api.get('core/version', null, false, { disableJWTAuthentication: true });
+    this.server_infos = (await Api.get('/internal/services/general/v1/server', null, false, {
+      disableJWTAuthentication: true,
+    })) as ServerInfoType;
 
-    if (!res.data) {
-      res.data = {};
-    }
-    this.server_infos = res.data;
-    this.server_infos.branding = this.server_infos.branding ? this.server_infos.branding : {};
     this.server_infos_loaded = true;
 
-    if (this.server_infos.ready !== true && this.server_infos.ready !== undefined) {
-      //Server is not ready
-      RouterServices.replace(RouterServices.addRedirection(RouterServices.pathnames.SETUP), {
-        ...this.server_infos.ready,
-      });
+    if (this.server_infos?.status !== 'ready') {
       this.app_ready = false;
       this.notify();
       setTimeout(() => {
