+++ twake/frontend/src/app/services/login/LoginService.ts
@@ -1,0 +1,5 @@
+import Observable from '../Observable/Observable';
+
+class LoginService extends Observable {}
+
+export default new LoginService();
